pyproject.toml
[project]
name = "tele-friend-bot"
version = "0.1.0"
description = "This is a simple telegram chatbot that will sit in a group chat and act like another friend."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "google-adk>=1.2.1",
    "litellm>=1.72.6",
    "psycopg2-binary>=2.9.10",
    "telethon>=1.34.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bot", "agentConversation"]
